requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiohttp-socks==0.8.0
aiosignal
h5py
anyio
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-timeout
attrs
backcall
beautifulsoup4
bleach
cchardet
certifi
cffi
charset-normalizer
click
colorama
comm
cycler
debugpy
decorator
defusedxml
demoji
gunicorn == 20.1.0
elastic-transport
elasticsearch
executing
flask
fake-useragent
fastjsonschema
fonttools
fqdn
frozenlist
geographiclib
geopy
googletransx
idna
imbalanced-learn
imblearn
ipykernel
ipython
ipython-genutils
ipywidgets
jedi
Jinja2
joblib
gcsfs
jsonpointer
jsonschema
jupyter
jupyter-console
jupyter_client
jupyter_core
jupyter_server
matplotlib
google-cloud-storage
matplotlib-inline
nltk
numpy
opendatasets
packaging
pandas
Pillow
PyYAML
regex
requests
scikit-learn
scipy
seaborn
tornado
tqdm
tensorflow
traitlets
twint
unzip
uri-template
wcwidth
webcolors
webencodings
websocket-client
widgetsnbextension
wordcloud